[core] make TransactionContexts reusable for subsequent commits
Support CodeSystemURI in all branch path based requests
Deprecate old repositoryId + branch request build methods.
Add throws declaration to LockRequest.doExecute(C).
