+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/request/BranchRequestBuilder.java
@@ -15,9 +15,13 @@
  */
 package com.b2international.snowowl.core.request;
 
+import com.b2international.snowowl.core.branch.Branch;
 import com.b2international.snowowl.core.domain.BranchContext;
 import com.b2international.snowowl.core.events.AsyncRequest;
+import com.b2international.snowowl.core.events.Request;
 import com.b2international.snowowl.core.events.RequestBuilder;
+import com.b2international.snowowl.core.uri.CodeSystemURI;
+import com.google.common.base.Strings;
 
 /**
  * Provides a default method for wrapping {@link BranchContext}-based requests
@@ -29,14 +33,44 @@ import com.b2international.snowowl.core.events.RequestBuilder;
  */
 public interface BranchRequestBuilder<R> extends RequestBuilder<BranchContext, R>, AllowedHealthStates {
 
+	/**
+	 * @param repositoryId
+	 * @param branch
+	 * @return
+	 * @deprecated - use {@link #build(String)} or {@link #build(CodeSystemURI)}, this method will be removed in 8.0
+	 */
 	default AsyncRequest<R> build(String repositoryId, String branch) {
-		return new AsyncRequest<>(
-			new RepositoryRequest<>(repositoryId,
-				new HealthCheckingRequest<>(
-					new BranchRequest<>(branch, build()),
-					allowedHealthstates()
+		// if the branch starts with MAIN, then it is an explicit branch path with a repositoryId
+		if (Strings.nullToEmpty(branch).startsWith(Branch.MAIN_PATH)) {
+			return new AsyncRequest<>(
+				new RepositoryRequest<>(repositoryId,
+					new HealthCheckingRequest<>(
+						new BranchRequest<>(branch, 
+							wrap(build())
+						),
+						allowedHealthstates()
+					)
 				)
+			);
+		} else {
+			return build(branch);
+		}
+	}
+
+	default AsyncRequest<R> build(String codeSystemUri) {
+		return build(new CodeSystemURI(codeSystemUri));
+	}
+	
+	default AsyncRequest<R> build(CodeSystemURI codeSystemUri) {
+		return new AsyncRequest<>(
+			new CodeSystemResourceRequest<>(
+				codeSystemUri,
+				wrap(build())
 			)
 		);
+	}
+	
+	default Request<BranchContext, R> wrap(Request<BranchContext, R> req) {
+		return req;
 	}
 }
