+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/request/RepositoryCommitRequestBuilder.java
@@ -83,7 +83,10 @@ public class RepositoryCommitRequestBuilder extends BaseRequestBuilder<Repositor
 
 	@Override
 	protected final Request<BranchContext, CommitResult> doBuild() {
-		return new TransactionalRequest(author, commitComment, getBody(), preparationTime, parentContextDescription);
+		return new RevisionIndexReadRequest<CommitResult>(
+			new TransactionalRequest(author, commitComment, getBody(), preparationTime, parentContextDescription),
+			false /*disable snapshot behaviour to allow the proper reuse of transaction context instances after a commit*/
+		);
 	}
 
 }
